ASpell uses AUnitSpell now
This removes duplicated code. ASpell adapts AUnitSpell.
AUnitSpell is the generic struct for custom unit spells.

The adapter pattern is used to adapt the structs in two directions.

ASpell has to always to create a channel trigger to check if the casting
unit is the character.

Static methods are part of AUnitSpell now.